Modifications
Modified Cell class to implement a 2D array for each cell (3x3).  Default constructors creates a 3x3 with '.'.  Overloaded creates a 3x3 with appropriate id in element 1,1.  Added a PrintArray() function that calls a ReturnCellValue from the Cell object defined inthe function.  This is the only way to use cout and should be removed if that ceases to be a requiremnt in the future.  There is quite a high cost for this function.