Pre-Final
Finished the program.  Dirty, so need some attention to move functions into their proper place.  the data has to be int instead of being able to parse a string.  I ran out of time.  Put in issues to address all of this so we cna fix and not have to mess with the software too much over the course.