Double Buffered Output
Preliminary function that uses a string copy of the console output and a copy of the previous output to double buffer and distinguish the changes.  The function then calls a setCursorPosition(x,y) function that puts the cursor at the change and allow the new value to overwrite the old value.  This prevents complete rewrite of the console.